move hook adjacent to related component
